fix filter

+++ src/main/java/cz/librarius/repository/dao/ListingRepoImpl.java
@@ -24,7 +24,7 @@ public class ListingRepoImpl extends GenericRepository<Listing> implements Listi
     }
 
     @Override
-    public List<Listing> findByFilter(Long isbn, String title, String author) {
+    public List<Listing> findByFilter(Double isbn, String title, String author) {
         TypedQuery<Listing> listingTypedQuery = em.createNamedQuery(Listing.FIND_BY_FILTER, Listing.class)
             .setParameter("isbn", isbn)
             .setParameter("title", title)
